Crate pallet_static_price_provider

Static Price Provider Pallet

A simple pallet that provides a static price for the native currency. This is a temporary solution before oracle is implemented & operational.

Overview

The Static Price Provider pallet provides functionality for setting the active native currency price via privileged call. Only the root can set the price.

Network maintainers must ensure to update the price at appropriate times so that inflation & dApp Staking rewards are calculated correctly.

Re-exports

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.